pyproject.toml
[project]
name = "testproject"
version = "0.1.0"
description = "FastAPI project testproject"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "sqlalchemy>=2.0.0",
    "psycopg[binary]>=3.1.12",
    "pydantic-settings>=2.8.1",
    "sqlmodel>=0.0.24",
    "asyncpg>=0.30.0",
    "psycopg2-binary>=2.9.10",
    "sqladmin[full]>=0.20.1",
    "alembic>=1.15.2",
    "uvicorn>=0.34.0",
    "fastapi-users[sqlalchemy]>=14.0.1",
    "fastapi-mail>=1.4.2",
    "fastapi-admin-cli>=0.1.13",

]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.0",
    "black>=23.9.1",
    "isort>=5.12.0",
    "ruff>=0.0.292",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
only-packages = true

[tool.hatch.build.targets.sdist]
include = [
    "app",
    "README.md",
]

[tool.hatch.build]
packages = ["app"]
